pyproject.toml
[project]
name = "langchain-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "pandas>=2.3.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "unstructured>=0.18.15",
    "wikipedia>=1.4.0",
]
